[Primitives - Custom Vertex Element, Cube , Refactoring]
Created a struct to represent a custom vertext element that can use position, normal, uv, and colour data per vertex.
This vertext struct can be used with the VertexBuffer and Dynamic VertexBuffer classes of XNA/Mono.
Added a cube primitive class that handles the creation of a cube. The cube uses Triangle List as a way to render the cube.

Refactored the names to better represent what the meshes are. Considering they are currently using VertexBuffer class they are mainly designed to be static and not modified at runtime.
Later when a Dynamic variant is introduced it will be distinguishable compared to the static variant at least.