[Primitives - More refactoring, Textured Cube]
Refactored all code from the game class into a separate class to allow for multiple 'scene' testing.
Corrected the 'normal' variable, accidentally made it a Vector2 instead of 3. Updated the vertex decleration accordingly to reflect the size difference.

Started implementing a textured cube. Realising that it carries a 'has a' relationship rather than an 'is a' relationship, it is not inherited.
I might refactor the other mesh classes to reflect this and see if it makes sense after refactoring.