[Primitives - Lighting Fixed]
Found the lighting issue. It was caused by the Custom Vertex Data element incorrectly defining the data type of the normal component as a Vector 2 instead of a Vector3.
Lighting works correctly and explains why the 'Z' axis was ignored for normal information, since it was told the component is a Vector2, so it only scans 8 bytes for the 2 float values instead of 12 bytes to scan for 3.
All works now, just need to create a scene to demo camera manipulation.