[Primitives - Transformation]
Now can use the mouse to transform the cube rendered in the scene.
Holding left mouse button and moving will rotate the cube.
Scrolling the mouse wheel will scale the cube.

Extra:
Added in the time class to this project from the previous 'MovingTextDemo' to track deltaTime.
Organised the classes into folders for less clutter.