[FPS Movement - Quaternions, Skybox, Camera controls]
Plane Object:
Created a plane game object that is able to render a mesh, be positioned, rotated, and scaled correctly.
Scale of this object is always done in local co-ordinates.

Quaternions:
Updated the camera object to allow for quaternion rotations.
Added directional vectors to the camera object.
Experimented with matrix world and view for the plane to see how the view and world matrix effect each other.
  --Seems that the world and view matrix are multiplied to each other. So the world can be used as a means of 'parenting' and object to another one's co-ordinate system.
  --And the view can be used to orient the object relative to 'a' view. The camera's or some other object.
  --Still want to experiment more with this.
Most of the functional examples of using quaternions are within the Camera class. This functionality will become the basis for a simple transform component for later objects.

Skybox:
Added a skybox model and learned how to draw the skybox to follow the player.
The skybox simply disables the depth buffer, renders itself at the position of the camera, and then re-enables the depth buffer after it is rendered.

Camera:
Camera controls have been made as an example of movement. Allowing the camera to move forward, back, left, and right relative to it's orientation.
Added controls for the camera to be able to look left, right, up, and down with the mouse's movement.