[FPS Movement - Jumping]
Added the ability to jump when the space bar is pressed.
Jumping is done via velocity and gravity is applied to the camera.
The velocity is linearly applied and slowed down.
The camera has been updated to remove a lot of Matrix recalculations and move them into 1 update function that checks if it is needed to update the view matrix.
Cleaned up some code a bit and added regions in the code for better readability.