[Raycasting and Projection - Basic Ray Projection]
A basic use of ray casting was created to test how projection is used to create the ray.
Also added a basic bounding box for the plane as well so the ray cast can check for intersection with it.

The code needs refactoring to allow for bounding volumes to be able to be added to an object and have it's points transformed according to it's matrices.
After playing with raycasting I will be starting on attempting to implement a crude component system.

Ray cast result is displayed in console window on left mouse click.