[Raycasting and Projection - Raycast Class]
Created a raycast class that can take the mouse position, and a bounding box and check if an intersection occurs.
If an intersection occurs it will return true and give out a struct containing the distance of the point plus the world position of the contact point.
Currently has a cube that will follow where the mouse has clicked on the ground.