[Raycast and Projection - MeshRendererComponent]
Added a first attempt at implementing a crude component using the built-in XNA GameComponent hierarchy.
It has a lot of oddities to it unfortunately but that is also due to my rushed attempt at seeing it work.
Currently there is a GameObject class as well that contains a transform component. It automates adding it's components into the main component collection of XNA.
Created a static mesh class to be used with the mesh renderer component. Need to add helper functions to construct basic primitives.
Transform class was missing scale vector.
Camera class has a static access variable for the main camera of the scene, which is the first camera to be spawned.
TexturedCube had it's mesh given public access for the sake of testing the GameObject cube for now, until a helper class is created. After which the TexturedCube class will be deleted.

Note:
Since an XNA GameComponent/Drawable requires the Game instance to be passed in from the constructor, it causes very cumbersome issues when making the 'AddComponent' method of GameObject a factory method.
Moving to a more Entity Component System might prove beneficial to circumvent this. Or simply passing the Component in as an argument, leaving the user responsible for constructing a component might be feasible for now.

Creating a singleton of the gameInstance might prove to be a very fast solution however, after reading about game programming patterns, I'd like to attempt avoiding singleton if possible.