[Raycasting and Projection - Components and Animated Tank]
Refactored a lot of the code now to introduce component based pattern within the project.
I'll break what's happened into sections.

Components:
Components have now been made using the Interface classes instead of inheriting the XNA GameComponent or DrawableGameComponent class. This was because having to pass the 'Game' instance to every component made it difficult to instantiate in a generic/factory manner.
Components have 2 types, a RenderComponent or a simple Component. A RenderComponent has a draw method for rendering and an Update method. A Regular component just has an update method.
Components also have a reference to their owner, which is a GameObject.

GameObject:
GameObjects now can have components added and removed from them. Currently addComponent has been tested and works as well as getComponent. RemoveComponent and the array variants of the generic functions haven't been tested yet.
GameObjects are responsible for adding their components into the XNA Components collection, however they also maintain their own list of components. Removing their components is also their responsibility.

Camera:
Camera has been updated to be a component now. All it's transformation stuff has been migrated into the transform.
Camera also has a static field called 'mainCamera' now that allows access to the first active camera spawned.

Plane:
Removed now. Is not necessary now that there is a helper class to construct meshes for the MeshRendererComponent.

PrimitiveShape:
A helper class used to construct basic primitive meshes to be used for renderering.

Game example:
The current game implemented in the project has a tank, which is a loaded model that can animate.
The tank can move thanks to a MoveToComponent and it's AnimatedTankMover.
While the tank moves, it animates.
The tank moves thanks to the AnimatedTankMover. It checks for a mouse click and casts a ray to find a contact point on a volume specified. Then it sets the destination of the MoveToComponent to this point.