[Raycasting and Projection - Clean up and Steering Behaviour]
Modified the MoveToComponent to be able to steer towards the destination as it moves as well as snap it's rotation to it as well.
Moved all functionality from the main Game class and instead created a player component that handles the player's interactions.
Moved the input functionality from AnimatedTankMover into the player component.
Created the player component which is pretty much just what I had in the game class. The name of the component isn't entirely accurate but it is a component not intended to be recycled.