[Arrive and Pursue - Base Project]
Added the base project for Arrive and Pursue behaviour experimentation.
Brought over the classes I implemented via the component game pattern from Raycasting and Projection and placed them into this project.
Also brought over the utility classes created too, with the game specific components from last project to start tweaking them a little.
Cleaned up namespaces of the classes and components. Base namespace is now the name of the repository.
Re-organised the classes into appropriate folders for this project. After this experimentation project, I will be proceeding to build somewhat of a game with what I have. While building the game I will be building on the codebase I have to implement a small engine with the hopes of adding an editor.

Aside from this, cleaned up namespacing in all files.