[Arrive and Pursue - Adding Arrive behaviour]
Created a component that creates the 'Arrive' behaviour for an object moving to a destination.
Refactored the MoveToComponent to be a simple lookAt and move forward component again.
Moved all gradual steering to be apart of the ArriveAtComponent.
Updated references to ArriveAtComponent where MoveTo was being used instead. Mainly in AnimatedTankMover.
Started on placing a clone function in the GameObject class however, I need to look more into ways of cloning objects. Either copy constructor or serialisation methods are currently what I am looking into.

Also updated readme to better explain the purpose of this repository.