[Arrive and Pursue - Pursue and Method Broadcasting]
Created a PursueComponent that at the moment does not inherit from ArriveAtComponent. Thought it seems using that might be a better idea but I will concern myself with that later when I am refactoring. PursueComponent now chases a target based on the targets trajectory of movement. It does not steal the target's destination and move to there, rather it tries to move to the point in from of the target's heading.

Also did rudimentary collision detection with the built in BoundingBox struct of XNA/MonoGame.
I have experimented with creating a cache of collided objects and broadcasting the 'OnCollision' method to the gameObject when a new collision occurs.
When the collision is no longer there, the object is removed from the collision cache and the 'OnCollisionExit' method is broadcasted to the gameObject.

Would like to experiment with this after I have refactored the pursuit code and gotten it working nicely.