[Arrive and Pursue - Collision Detector Service]
Added a Collision Detector service that currently simply performs the O(n2) collision detection scheme.
The collision detection keeps a cache of collided objects and registers a collision event depending on if there is a new collision or a stopped collision.
2 messages are sent by the collision detector, OnCollision and OnCollisionExit.

The pursue component was updated to use the OnCollision function to stop itself and the OnCollisionExit function to continue movement.

A Collider Component was made however it is currently just a base class sitting there until I find common functionality between sphere and box bounding volumes with XNA/Monogame.
A BoxCollider component was made which takes a bounding box and allows for collision intersections to be checked for with other bounding boxes.