[Arrive and Pursue - Refactoring]
Refactored the PursueComponent to inherit from ArriveAtComponent.
PursueComponent now will have a cooldown once collision with the target has stopped. When the cooldown is over, it will start pursuing again.
Made a duplicate of the tank model to be used by the second tank(pursuit one) so it can have it's own animations playing.
    -This is from the fact that model's are instanced. Because the AnimatedTank component modifies the model's bones, it is reflected in the second tank object.
    -So I just made a 2nd tank fbx as a workaround for now.
Cleaned up leftover comments in some files.
Removed Destination property in the AnimatedTankMover. It was redundant and unnecessary.