[Maze Escape - Base Project]
Created a new project to experiment with creating a small game with the current code base.
Base Components were moved over to the project as well as utilities from Arrive and Pursue project.

The basis of this project will be to test 3D level creation and then later maybe some path finding.
Also to hopefully start on some physics implementations and make a 'rigidbody' component.
The rigidbody component will be able to have physics applied to it and simulated in fixed time steps as well as have collision resolutions with other colliders.
   -Main purpose of the rigidbody will be to have force application, collision response, and fixed simulation steps.
   -Angular forces will not be focused on nor will contact points in collision response early on.