[Maze Escape - Level Loading and Input Utility]
Created a level loader which simply reads in a text file and parses the text file into data structure.
The level loader will find all level files within a 'Levels' folder relative to the executable path.
After loading all level files, a level can be parsed by specifying a file name.
The LevelData structure stores the level as a flat array of ints with helper functions to access the array.
The Game class has a basic showing of the level loader working.

Refactored the Time class to allow for static access, instead of constantly having to use .Instance.
Created a utility class for Input, which handles common input functionality with the keyboard and mouse.