[Maze Escape - Alpha level + hazards]
Added hazards into the level construction which use the pursue behaviour on the player.
Added pickups to the game which are merely yellow gameobjects.
The player's firstpersoncontroller component keeps track of score and can play sound effects now for this alpha.
CollisionDetector re-written due to crippling performance when I tried to correct a bug with collision caching.
Now collision detector has a dynamic colliders list and static colliders list.