[Pathfinding - Node Creation]
Added in a very rushed implementation of constructing a level graph for pathfinding to be used on.
A simple scene is made which currently just allows the player to click on the ground and move to that point.
Later when the ground is clicked, that point will be found in the level graph, and then a path from the player's point to it will be found via an algorithm.