[Finite State Machine - Base Project]
Setup of base project for finite state machine. Will be used to test creating a finite state machine as well as having it be able to use an external file for configuration and authoring of different AI behaviours.
Migrated over code base from Pathfinding project, with some refactoring of namespaces.