[Finite State Machine - FSM XML File parsing]
Added an NPC cube that for now just wanders around via pathfinding.
Created an interpreter for the XML file that will parse it into a structure of machines that have states.
Every state has transitions, and these transitions have a target state to go to and an evaluation condition.

Currently investigating how to encapsulate and self contain a state's execution function as well as a transitions conditional evaluation.
Thinking of using the command pattern via C#'s built in Func delegate type.