[Finite State Machine - AI Processing]
After FSMInterpreter has parsed the XML file, it then passes on the 'activeFSM' to the NPCController which will setup it's machine by giving task delegations to each state and transition conditional expressions to each state's transition.
Currently the AI can perform it's machine processes however the pathfinding is a bit unresponsive or rather inaccurate when having to constantly update it's starting node.
The definition of the AI behaviour can be seen in the xml file 'fsm_npc1.xml'.

There is a power up for the player that can be collected, if it is collected the player is powered and will cause the NPC to be destroyed in the event of a collision.
Otherwise the player is destroyed instead in the event of a collision.

Removed FSMComponent as a use case for it was made redundant during testing.