[Maze Escape - Code update and migration]
Updated and migrated code from FSM project into the Maze Escape project. The updates were minor, however this was done to unify the code base for Maze Escape to easily reflect and use latest versions of the code for the final test of Maze Escape which will have NPCs use pathfinding for movement and include a rudimentary BSP implementation for collisions.