[Maze Escape - Destroyable objects and nullification]
Began work on nullifying a gameobject once it is destroyed as well as implementing a way to destroy gameObjects 'cleanly'.
Also discovered that components added at runtime have their initialisation function called immediately. Might have to change how things work in Game class and components to reflect this call pipeline to make it easier to interface and create components and objects at runtime.