[Maze Escape - Game1 class cleanup]
Created a 'Scene' class which contains a 'game' or more thought of as a level that can be loaded by the game.
The game class was modified so that it would initialise all it's base properties, and then within it's update in the 1st frame it would then initialise the scene.
This was done because between the game initialising and the first update frame there is a window of time components added will not be initialised.

Moved all stuff inside of the Game1 class into it's own component or into the MazeScene class.

Some component classes were cleaned up of unnecessary lines and modified some getters to reflect the new framework initialisation order. Now components can load their own content!