[Maze Escape - Text Display Component]
Created a component that can render text to the screen, with a position and colour associated with the component.
Also moved out some of the code from MazeScene into a 'LevelBuilder' which handles the parsing and instantiation of tiles from a level.
The Maze Scene uses the parses points from the level builder to then populate the scene with enemies, collectables, and the player.